[FIX] use git to update MACS (#1228)
* use git to update MACS

* Update bidspm.m


* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Apply suggestions from code review

---------

+++ src/infra/checkToolbox.m
@@ -65,7 +65,7 @@ function status = checkToolbox(varargin)
 
       if ~status && install
 
-        installMacstoolbox(opt);
+        updateMacstoolbox();
 
         status = checkToolbox(toolboxName);
 
@@ -89,19 +89,27 @@ function status = checkToolbox(varargin)
 
 end
 
-function installMacstoolbox(opt)
+function updateMacstoolbox()
+
   SPM_DIR = spm('dir');
-  MACS_TOOLBOX_DIR = fullfile(returnRootDir(), 'lib', 'MACS');
-
   target_dir = fullfile(SPM_DIR, 'toolbox', 'MACS');
+  MACS_TOOLBOX_DIR = fullfile(rootDir(), 'lib', 'MACS');
 
-  msg = sprintf('installing MACS toolbox in:\n%s.\n\n', target_dir);
-  id = 'installingMacsToolbox';
-  logger('WARNING', msg, 'id', id, 'filename', mfilename(), 'options', opt);
-
-  if exist(target_dir, 'dir') == 7
+  if exist(target_dir, 'dir') == 7 && exist(fullfile(target_dir, '.git'), 'dir') == 0
     rmdir(target_dir, 's');
   end
-  mkdir(target_dir);
-  copyfile(MACS_TOOLBOX_DIR, target_dir);
+
+  if exist(target_dir, 'dir') == 7
+    msg = sprintf('updating MACS toolbox: ');
+    fprintf(1, msg);
+    system(sprintf('git -C %s pull', ...
+                   target_dir));
+
+  else
+    msg = sprintf('installing MACS toolbox in:\n%s.\n', target_dir);
+    fprintf(1, msg);
+    system(sprintf('git clone %s %s', ...
+                   MACS_TOOLBOX_DIR, ...
+                   target_dir));
+  end
 end
